Moved CutsceneArea to ui/cutscene/
